requirements_test.txt
pytest-homeassistant-custom-component>=0.8.11
aiohttp_cors>=0.7.0
aiodiscover>=1.4.8
scapy>=2.4.5
async_upnp_client>=0.27.0
pyheif>=0.7.0
Pillow>=9.2.0

codecov>=2.1.12
coverage>=6.3.2
pytest>=7.1.1
pytest-cov>=2.12.1
pytest-httpx>=0.20.0